CoST:
Change the length of the control measures to 20 from 10.

EMF:

1.	Bug 3621: Bug 3621: list of key features of each case

Added ÒDescribeÓ button to bottom of Case Window (both the Editor and Viewer Windows), which brings up non-modal editor/viewer window when it is clicked. For the one invoked from the Editor window, make the test field for the description edtitable, and the updated description can be transferred to Case Editor window and cab be saved.

2.	Bug 3615: Delete QA step

Fixed. Dependences also checked before deletion; result table also deleted.

3.	Bug 3588: SQL query to detect bizarre characters in plant name columns

Fixed.
1). If there are exceptions when viewing QA step results, check if it is caused by the bizarre characters, if yes, report error and remind the use to run QA step Detect Bizarre Characters in Plant Name, otherwise report error.
2). Use Regular Expression to detect the bizarre characters Ð the control characters (ASCII 0-31 and 127, if also include Unicode 128-65535, it becomes slow Ð do we need to detect them too?)
3). Do we need to develop a solution to remove the bizarre characters? A SQL script to do the update is ready.

4.	Bug 3516: Externalize Server Debug Level

Fixed. Debug level externalized. The static final DEBUG_xx change to a function, such that it can query the database properties table dynamically.
Further consideration:
1). Override log4j class to a more powerful logger to control the logged messages?
2). Specify horizontal and vertical debugging levels? Horizontal: CoST, CASE, etc. Vertical: e.g. 0 Ð very few debug information, 10 Ð very detailed information.

5.	Bug 3618: Issue with QA step EMF Extended SQL Syntax.

Invested the problem and need to fix that later according to the discussion. Need to check the EMF extended sql parser for the following patterns: need to check all the following patterns:
$TABLE [table_num] a
$DATASET_TABLE [ Òdataset nameÓ, table_num] a
$DATASET_TABLE_VERSION [ Òdataset nameÓ, table_num, version_num] a
$DATASET_QASTEP [ Òdataset nameÓ, ÒQA step nameÓ] a
$DATASET_QASTEP_VERSION [Òdataset nameÓ, ÒQA step nameÓ, version_num]